validator: give the disruptive proxy groups a cluster of their own
Restarting nodes, cutting their traffic off and restarting Vector Store
with a different configuration is more than a sibling group can survive,
so these groups keep a cluster to themselves. They still take
the session, keyspace and teardown from the same fixture as everyone
else.
