+++ crates/validator/src/owned/connection_timeout.rs
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.1
  */
 
-use crate::TestActors;
+use super::OwnedProxyCluster;
+use super::OwnedProxyContext;
 use crate::common::*;
 use e2etest_firewall::FirewallExt;
 use e2etest_vector_store_cluster::VectorStoreClusterExt;
@@ -16,25 +17,9 @@ const CONNECTION_TIMEOUT: &str = "5s";
 
 e2etest::group!(
     name = connection_timeout,
-    fixtures = (Fixture),
+    fixtures = (OwnedProxyCluster),
     parent = super::owned
 );
-
-struct Fixture {
-    actors: Arc<TestActors>,
-}
-
-impl e2etest::Fixture for Fixture {
-    async fn setup(setup: &mut impl e2etest::Setup) -> Option<Self> {
-        let actors = setup.setup::<TestActors>().await?;
-        init_with_proxy_single_vs(&actors).await;
-        Some(Self { actors })
-    }
-
-    async fn teardown(self) {
-        cleanup(&self.actors).await;
-    }
-}
 
 /// Test that the CQL connection timeout causes session creation to fail
 /// when the database is unreachable, and that vector-store recovers once
@@ -49,8 +34,10 @@ impl e2etest::Fixture for Fixture {
 /// - Restore connectivity by allowing traffic through the proxy.
 /// - Verify that vector-store eventually connects and becomes ready.
 #[e2etest::test(group = connection_timeout)]
-async fn connection_timeout_triggers_session_failure(actors: Arc<TestActors>) {
+async fn connection_timeout_triggers_session_failure(ctx: Arc<OwnedProxyContext>) {
     info!("started");
+
+    let actors = ctx.actors();
 
     info!("Stop vector-store");
     actors.vs.stop().await;
@@ -58,14 +45,14 @@ async fn connection_timeout_triggers_session_failure(actors: Arc<TestActors>) {
     info!("Block all traffic through the proxy");
     actors
         .firewall
-        .drop_traffic(get_default_db_proxy_ips(&actors))
+        .drop_traffic(get_default_db_proxy_ips(actors))
         .await;
 
     info!("Restart vector-store with CQL connection timeout");
     actors
         .vs
-        .start(get_proxy_vs_node_configs(&actors).pipe(|mut nodes| {
-            let translation_map = get_proxy_translation_map(&actors);
+        .start(get_proxy_vs_node_configs(actors).pipe(|mut nodes| {
+            let translation_map = get_proxy_translation_map(actors);
             for node in nodes.iter_mut() {
                 node.envs.insert(
                     "VECTOR_STORE_CQL_URI_TRANSLATION_MAP".to_string(),
@@ -82,7 +69,7 @@ async fn connection_timeout_triggers_session_failure(actors: Arc<TestActors>) {
         .await;
 
     info!("Wait for VS HTTP to become reachable");
-    let vs_ips = get_default_vs_ips(&actors);
+    let vs_ips = get_default_vs_ips(actors);
     let vs_addr = std::net::SocketAddr::from((vs_ips[0], VS_PORT));
     let client = httpclient::HttpClient::new(vs_addr);
     wait_for(
